Red
Red Aura people are enthusiastic and energetic individuals, forever on
the lookout for new adventures. They are adventurous with food, travel
and sexual partners. The mantra of the Red Aura colour individual is
"I'll try anything once." Because of their devil-may-care approach to
life they often find themselves in hot water.